first_insight/requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.2
idna==2.6
lxml==3.8.0
Markdown==2.6.9
mysqlclient==1.3.11
pytz==2017.2
requests==2.18.4
selenium==3.5.0
urllib3==1.22
urlopen==1.0.0
wheel==0.29.0
